refactor: split scaler into per-algorithm modules
Move each scaling algorithm into its own file under standalone/scaler/scalers/:

- bilinear.rs: 2x2 neighbourhood with fixed-point weights

- bicubic.rs: separable two-pass Catmull-Rom

- xbrz.rs: xBRZ pixel-art scaler

- mod.rs: shared types (BiAxisMap, BcAxisMap) and axis-map builders

The dispatcher (scaler/mod.rs) holds per-algorithm state and routes

scale() calls based on the active ScaleFilter variant.

Closes #61
