+++ crates/native32emu/src/standalone/mod.rs
@@ -3,4 +3,3 @@
 pub mod cli;
 pub mod gamepad_overlay;
 pub mod scaler;
-pub mod xbrz;
